Add react and restore legacy functionality now being served by rails without gatsby
