+++ app/controllers/site_controller.rb
@@ -1,0 +1,4 @@
+class SiteController < ApplicationController
+  def index
+  end
+end
