+++ app/javascript/components/footer.js
@@ -2,11 +2,13 @@ import React from 'react';
 
 import '../styles/footer.css';
 
-const Footer = ({ children }) => (
+const Footer = () => (
   <footer>
     © {new Date().getFullYear()}, Built with ❤️ using
     {` `}
-    <a href="https://www.gatsbyjs.org">Gatsby</a>
+    <a href="https://rubyonrails.org/">Rails 6</a>
+    {` & `}
+    <a href="https://reactjs.org/">React</a>
   </footer>
 )
