+++ app/javascript/components/slackInviteForm.js
