+++ app/javascript/pages/index.js
@@ -1,13 +1,11 @@
 import React from 'react';
 
 import Layout from '../components/layout';
-import SEO from '../components/seo';
 import SlackInviteForm from '../components/slackInviteForm';
 import { SEO_TAGS, SEO_TITLES } from '../utils/constants';
 
 const IndexPage = () => (
   <Layout>
-    <SEO title={SEO_TITLES.homepage} keywords={SEO_TAGS} />
     <h1 className='main'>
       Calling All
       Los Angeles-based Developers
