+++ app/views/site/index.html.erb
@@ -1,0 +1,1 @@
+<div id="root"></div>
