+++ package.json
@@ -2,10 +2,15 @@
   "name": "ladevs_org",
   "private": true,
   "dependencies": {
+    "@babel/preset-react": "^7.0.0",
     "@rails/actioncable": "^6.0.0-alpha",
     "@rails/activestorage": "^6.0.0-alpha",
     "@rails/ujs": "^6.0.0-alpha",
     "@rails/webpacker": "^4.0.7",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "turbolinks": "^5.2.0"
   },
   "version": "0.1.0",
